[generator] move some code to separate files, renamings
