+++ generator/geo_objects/geo_objects.hpp
@@ -1,7 +1,10 @@
 #pragma once
 
 #include "generator/key_value_storage.hpp"
+
 #include "generator/regions/region_info_getter.hpp"
+
+#include "generator/geo_objects/geo_objects_maintainer.hpp"
 
 #include "geometry/meter.hpp"
 #include "geometry/point2d.hpp"
@@ -18,6 +21,34 @@ namespace generator
 {
 namespace geo_objects
 {
+class RegionInfoGetterProxy
+{
+public:
+  using RegionInfoGetter = std::function<boost::optional<KeyValue>(m2::PointD const & pathPoint)>;
+  RegionInfoGetterProxy(std::string const & pathInRegionsIndex,
+                        std::string const & pathInRegionsKv)
+  {
+    m_regionInfoGetter = regions::RegionInfoGetter(pathInRegionsIndex, pathInRegionsKv);
+    LOG(LINFO, ("Size of regions key-value storage:", m_regionInfoGetter->GetStorage().Size()));
+  }
+
+  explicit RegionInfoGetterProxy(RegionInfoGetter && regionInfoGetter)
+      : m_externalInfoGetter(std::move(regionInfoGetter))
+  {
+    LOG(LINFO, ("External regions info provided"));
+  }
+
+  boost::optional<KeyValue> FindDeepest(m2::PointD const & point) const
+  {
+    return m_regionInfoGetter ? m_regionInfoGetter->FindDeepest(point)
+                              : m_externalInfoGetter->operator()(point);
+  }
+private:
+  boost::optional<regions::RegionInfoGetter> m_regionInfoGetter;
+  boost::optional<RegionInfoGetter> m_externalInfoGetter;
+};
+
+
 using IndexReader = ReaderPtr<Reader>;
 
 boost::optional<indexer::GeoObjectsIndex<IndexReader>> MakeTempGeoObjectsIndex(
@@ -25,74 +56,35 @@ boost::optional<indexer::GeoObjectsIndex<IndexReader>> MakeTempGeoObjectsIndex(
 
 bool JsonHasBuilding(JsonValue const & json);
 
-class GeoObjectsGenerator
+void AddBuildingsAndThingsWithHousesThenEnrichAllWithRegionAddresses(
+    KeyValueStorage & geoObjectsKv, RegionInfoGetterProxy const & regionInfoGetter,
+    std::string const & pathInGeoObjectsTmpMwm, bool verbose, size_t threadsCount);
+
+struct NullBuildingsInfo
 {
-public:
-  using RegionInfoGetter = std::function<boost::optional<KeyValue>(m2::PointD const & pathPoint)>;
-
-  class RegionInfoGetterProxy
-  {
-  public:
-    RegionInfoGetterProxy(std::string const & pathInRegionsIndex,
-                          std::string const & pathInRegionsKv)
-    {
-      m_regionInfoGetter = regions::RegionInfoGetter(pathInRegionsIndex, pathInRegionsKv);
-      LOG(LINFO, ("Size of regions key-value storage:", m_regionInfoGetter->GetStorage().Size()));
-    }
-
-    explicit RegionInfoGetterProxy(RegionInfoGetter && regionInfoGetter)
-      : m_externalInfoGetter(std::move(regionInfoGetter))
-    {
-      LOG(LINFO, ("External regions info provided"));
-    }
-
-    boost::optional<KeyValue> FindDeepest(m2::PointD const & point) const
-    {
-      return m_regionInfoGetter ? m_regionInfoGetter->FindDeepest(point)
-                                : m_externalInfoGetter->operator()(point);
-    }
-  private:
-    boost::optional<regions::RegionInfoGetter> m_regionInfoGetter;
-    boost::optional<RegionInfoGetter> m_externalInfoGetter;
-  };
-
-  GeoObjectsGenerator(std::string pathInRegionsIndex, std::string pathInRegionsKv,
-                      std::string pathInGeoObjectsTmpMwm, std::string pathOutIdsWithoutAddress,
-                      std::string pathOutGeoObjectsKv, bool verbose, size_t threadsCount);
-
-  GeoObjectsGenerator(RegionInfoGetter && regionInfoGetter, std::string pathInGeoObjectsTmpMwm,
-                      std::string pathOutIdsWithoutAddress, std::string pathOutGeoObjectsKv,
-                      bool verbose, size_t threadsCount);
-
-  // This function generates key-value pairs for geo objects.
-  // First, we try to generate key-value pairs only for houses, since we cannot say anything about
-  // poi. In this step, we need key-value pairs for the regions and the index for the regions. Then
-  // we build an index for houses. And then we finish building key-value pairs for poi using this
-  // index for houses.
-  bool GenerateGeoObjects();
-
-  KeyValueStorage const & GetKeyValueStorage() const
-  {
-    return m_geoObjectsKv;
-  }
-
-private:
-  bool GenerateGeoObjectsPrivate();
-  static KeyValueStorage InitGeoObjectsKv(std::string const & pathOutGeoObjectsKv)
-  {
-    Platform().RemoveFileIfExists(pathOutGeoObjectsKv);
-    return KeyValueStorage(pathOutGeoObjectsKv, 0 /* cacheValuesCountLimit */);
-  }
-
-  std::string m_pathInGeoObjectsTmpMwm;
-  std::string m_pathOutPoiIdsToAddToLocalityIndex;
-  std::string m_pathOutGeoObjectsKv;
-
-  bool m_verbose = false;
-  size_t m_threadsCount = 1;
-
-  KeyValueStorage m_geoObjectsKv;
-  RegionInfoGetterProxy m_regionInfoGetter;
+  std::unordered_map<base::GeoObjectId, base::GeoObjectId> m_addressPoints2Buildings;
+  // Quite possible to have many points for one building. We want to use
+  // their addresses for POIs according to buildings and have no idea how to distinguish between
+  // them, so take one random
+  std::unordered_map<base::GeoObjectId, base::GeoObjectId> m_Buildings2AddressPoint;
 };
+
+NullBuildingsInfo EnrichPointsWithOuterBuildingGeometry(
+    GeoObjectMaintainer const & geoObjectMaintainer, std::string const & pathInGeoObjectsTmpMwm,
+    size_t threadsCount);
+
+
+void AddPoisEnrichedWithHouseAddresses(KeyValueStorage & geoObjectsKv,
+                                       GeoObjectMaintainer const & geoObjectMaintainer,
+                                       NullBuildingsInfo const & buildingsInfo,
+                                       std::string const & pathInGeoObjectsTmpMwm,
+                                       std::ostream & streamPoiIdsToAddToLocalityIndex,
+                                       bool verbose, size_t threadsCount);
+
+
+void FilterAddresslessThanGaveTheirGeometryToInnerPoints(std::string const & pathInGeoObjectsTmpMwm,
+                                                         NullBuildingsInfo const & buildingsInfo,
+                                                         size_t threadsCount);
+
 }  // namespace geo_objects
 }  // namespace generator
