+++ generator/geo_objects/geo_objects_maintainer.cpp
@@ -1,4 +1,4 @@
-#include "generator/geo_objects/geo_object_maintainer.hpp"
+#include "generator/geo_objects/geo_objects_maintainer.hpp"
 
 #include <utility>
